requirements.txt
# requirements.txt - Job Search Agent Dependencies

# Google API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
requests>=2.28.0
